+++ src/ts/beatmap/waypoint.ts
@@ -1,0 +1,8 @@
+// as far as i know, it does not have customData as of yet
+export interface Waypoint {
+    _time: number;
+    _lineIndex: number;
+    _lineLayer: number;
+    _offsetDirection: number;
+    [key: string]: number;
+}