+++ src/ts/tools/analyse.ts
@@ -2,10 +2,20 @@ import savedData from '../savedData';
 import { CharacteristicName } from '../beatmap/characteristic';
 import { DifficultyName } from '../beatmap/difficulty';
 
-const general = (): void => {};
-const difficulty = (mode?: CharacteristicName, difficulty?: DifficultyName): void => {
+interface Analysis {
+    missing: [];
+    mapSet: AnalysisSet;
+}
+
+interface AnalysisSet {
+    mode: CharacteristicName;
+    difficulty: DifficultyName;
+    sps: number;
+    htmlOutput: HTMLElement | string;
+}
+
+export const general = (): void => {};
+export const difficulty = (mode?: CharacteristicName, difficulty?: DifficultyName): void => {
     if (mode && difficulty) {
     }
 };
-
-export default { general, difficulty };