+++ src/ts/ui/theme.ts
@@ -1,7 +1,7 @@
 // theme should not contain space and will be converted to lowercase
 // TODO: maybe separate the theme into another folder/file?
 export type Theme = 'Dark' | 'Light' | 'Monochrome';
-export const themeList: Theme[] = ['Dark', 'Light', 'Monochrome'];
+export const list: ReadonlyArray<Theme> = ['Dark', 'Light', 'Monochrome'];
 
 const htmlBody = document.querySelector<HTMLBodyElement>('body');